Move Tropical into its own module.
