+++ fresnel/test/Test.hs
@@ -26,6 +26,7 @@ import           Fresnel.Lens (Lens', lens)
 import           Fresnel.Maybe (_Just)
 import           Fresnel.Optional (is)
 import           Fresnel.Setter
+import           Fresnel.Tropical
 import           GHC.Exception.Type (Exception(displayException))
 import qualified Getter.Test
 import qualified Iso.Test
